Update to work with the latest version of brackets.
The className has changed for css properties, and there is now a meta
class for the vendor prefix. Note that with this change, if the cursor
is placed right before the property name, the editor will no longer
appear.

Also fixed a minor bug with a , instead of a ;. JSLint yelled at me as
soon as I opened this up in Brackets. Thank you, JSLint.